Combine scripts into one script
