--- copy_protection/gen_keys_header.py
@@ -1,122 +1,0 @@
-#!/usr/bin/env python3
-
-import sys
-
-def is_unique_and_in_range(byte_array, min, max):
-    # Проверка диапазона для всех байт
-    in_range = all(min <= b <= max for b in byte_array)
-    # Проверка уникальности — множество хранит только уникальные значения
-    unique = len(set(byte_array)) == len(byte_array)
-    return in_range and unique
-
-
-def read_hmac_swap_table_from_file(filename):
-    arrays = []
-
-    with open(filename, "r") as f:
-        for line in f:
-            # Ignore commented lines
-            if line.startswith("#"):
-                continue
-            # Delete new lines and spaces
-            hex_str = line.strip()
-            if not hex_str:
-                continue
-            # Convert string to bytes
-            data = bytes.fromhex(hex_str)
-            arrays.append(data)
-
-    # Check lines count
-    if len(arrays) != 2:
-        raise ValueError(f"File '{filename}' must contain 2 strings with hex values")
-
-    hmac_key = arrays[0]
-    if len(hmac_key) != 32:
-        raise ValueError("HMAC key must be 32 bytes in length")
-
-    swap_table = arrays[1]
-    if len(swap_table) != 12:
-        raise ValueError("Swap table must be 12 bytes in length")
-    if not is_unique_and_in_range(swap_table, 0x00, 0x1F):
-        raise ValueError("Swap table contain only unique bytes in range 0x00...0x1F")
-
-    return hmac_key, swap_table
-
-
-def read_swap_tables_from_file(filename):
-    arrays = []
-
-    with open(filename, "r") as f:
-        for line in f:
-            # Ignore commented lines
-            if line.startswith("#"):
-                continue
-            # Delete new lines and spaces
-            hex_str = line.strip()
-            if not hex_str:
-                continue
-            # Convert string to bytes
-            data = bytes.fromhex(hex_str)
-            arrays.append(data)
-
-    # Check lines count
-    if len(arrays) != 2:
-        raise ValueError(f"File '{filename}' must contain 2 strings with hex values")
-
-    hmac_table = arrays[0]
-    if len(hmac_table) != 32:
-        raise ValueError("HMAC swap table must be 32 bytes in length")
-    if not is_unique_and_in_range(hmac_table, 0x00, 0x1F):
-        raise ValueError("HMAC swap table contain only unique bytes in range 0x00...0x1F")
-
-    swap_table = arrays[1]
-    if len(swap_table) != 12:
-        raise ValueError("Swap table must be 12 bytes in length")
-    if not is_unique_and_in_range(swap_table, 0x00, 0x0B):
-        raise ValueError("Swap table contain only unique bytes in range 0x00...0x0B")
-
-    return hmac_table, swap_table
-
-
-def generate_keys_header_file(filename, key, key_table, swap, swap_table):
-    with open(filename, "w") as f:
-        f.write("#pragma once\n\n")
-        f.write(f"// This file was generated by the {sys.argv[0]} script\n\n")
-
-        hex_values = ", ".join(f"0x{b:02X}" for b in key)
-        define_str = "#define HMAC_KEY"
-        spaces = " " * (32 - len(define_str))
-        f.write(f"{define_str}{spaces}{{{hex_values}}}\n")
-
-        hex_values = ", ".join(f"0x{b:02X}" for b in key_table)
-        define_str = "#define HMAC_KEY_TABLE"
-        spaces = " " * (32 - len(define_str))
-        f.write(f"{define_str}{spaces}{{{hex_values}}}\n\n")
-
-        hex_values = ", ".join(f"0x{b:02X}" for b in swap)
-        define_str = "#define PROT_CODE_SWAP"
-        spaces = " " * (32 - len(define_str))
-        f.write(f"{define_str}{spaces}{{{hex_values}}}\n")
-
-        hex_values = ", ".join(f"0x{b:02X}" for b in swap_table)
-        define_str = "#define PROT_CODE_SWAP_TABLE"
-        spaces = " " * (32 - len(define_str))
-        f.write(f"{define_str}{spaces}{{{hex_values}}}\n")
-
-        print(f"The keys were wtitten to the '{filename}' file")
-
-
-def apply_swap(values, swap_table):
-    result = bytearray(len(values))
-    for i, pos in enumerate(swap_table):
-        result[i] = values[pos]
-    return bytes(result)
-
-
-hmac_key, swap = read_hmac_swap_table_from_file("keys.txt")
-hmac_table, swap_table = read_swap_tables_from_file("swap_tables.txt")
-
-result_hmac_key = apply_swap(hmac_key, hmac_table)
-result_swap = apply_swap(swap, swap_table)
-
-generate_keys_header_file("../main/copy_protection/keys.h", result_hmac_key, hmac_table, result_swap, swap_table)
