+++ copy_protection/readme.md
@@ -13,7 +13,7 @@
 Сначала генерируется 32-байтный хеш из 6-байтного MAC-адреса с использованием 32-байтного ключа.
 Далее из 32-байтного хеша формируется 12-байтный код защиты с использованием таблицы перестановок.
 
-## Генератор кода защиты
+## Генерация кода защиты
 
 Ключи (ключ шифрования и таблица перестановок) заданы в файле `keys.txt`.
 Файл должен содержать 2 строки с 16-ричными значениями, разделенными пробелами.
@@ -23,11 +23,13 @@
 
 В файле ключей допускаются пустые строки и комментарии, начинающиеся с символа `#` в начале строки.
 
-Для генерации кода по MAC-адресу используется скрипт `gen_sec_code.py`.
+Для генерации кода по MAC-адресу используется скрипт `copy_prot_helper.py`.
 
 Пример использования:
+
 ``` bash
-./gen_sec_code.py --mac 04:83:08:D7:B3:00
+cd copy_protection
+./copy_prot_helper.py --mac 04:83:08:D7:B3:00 --keys keys.txt --out sec_code.bin
 ```
 В результате будет сгенерирован бинарный файл `sec_code.bin` с кодом защиты, размер файла 12 байт.
 Данный файл можно использовать для прошивки кода во фьюзы устройства с помощью утилиты `espefuse.py`.
@@ -42,9 +44,9 @@
 
 Данные массивы храняется не подряд, а с некоторыми произвольными отступами.
 
-## Генератор заголовочного файла с ключами
+## Генерация заголовочного файла с ключами
 
-При сборке проекта для генерации заголовочного файла с ключами используется скрипт `gen_keys_header.py`.
+При сборке проекта для генерации заголовочного файла с ключами используется этот же скрипт `copy_prot_helper.py`.
 
 На входе скрипт получает те же ключи из файла `keys.txt`, а также таблицы перестановок из файла `swap_tables.txt`.
 
@@ -56,7 +58,14 @@
 
 В файле таблиц перестановок `swap_tables.txt` допускаются пустые строки и комментарии, начинающиеся с символа `#` в начале строки.
 
-В результате работы скрипта `gen_keys_header.py` генерируется заголовочный файл `keys.h`, используемый при сборке проекта.
+Пример использования:
+
+``` bash
+cd copy_protection
+./copy_prot_helper.py --keys keys.txt --swap_tables swap_tables.txt --out_header ../main/copy_protection/keys.h
+```
+
+В результате работы скрипта `copy_prot_helper.py` генерируется заголовочный файл `keys.h`, используемый при сборке проекта.
 
 ## Логика проверки валидности устройства
